+++ CHANGELOG.md
@@ -8,6 +8,9 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 
 ## [Unreleased]
 
+### Added
+- `phylum auth list-tokens` subcommand to list API tokens
+
 ## [5.6.0] - 2023-08-08
 
 ### Added